+++ silo-core/contracts/lib/SiloSolvencyLib.sol
@@ -138,7 +138,7 @@ library SiloSolvencyLib {
 
         uint256 totalDebtAssets = _accrueInMemory == ISilo.AccrueInterestInMemory.Yes
             ? SiloStdLib.getTotalDebtAssetsWithInterest(_debtConfig.silo, _debtConfig.interestRateModel)
-            : ISilo(_debtConfig.silo).getDebtAssets();
+            : ISilo(_debtConfig.silo).total(ISilo.AssetType.Debt);
 
         // BORROW value -> to assets -> UP
         ltvData.borrowerDebtAssets = SiloMathLib.convertToAssets(