+++ silo-core/test/foundry/lib/SiloLiquidationLib/SiloLiquidationLib.t.sol
@@ -332,7 +332,7 @@ contract SiloLiquidationLibTest is Test, MaxRepayRawMath {
         uint128 _ltvBefore,
         uint128 _sumOfCollateralAssets,
         uint128 _debtToCover
-    ) public {
+    ) public pure {
         // total assets/values must be != 0, if they are not, then revert possible
         uint256 borrowerDebtAssets = 1e18;
         uint256 borrowerDebtValue = 1e18;